refactor(tools): migrate utility tools to directory structure
Refactored Date/Time Tools (5 tools), String Utilities (7 tools), Math Operations (4 tools), and Validation Tools (6 tools) into organized directory structures following the established pattern. All 975 tests passing.
